+++ Classes/Private/Manager/Configuration/OCTManagerConfiguration.m
@@ -7,10 +7,8 @@
 //
 
 #import "OCTManagerConfiguration.h"
-#import "OCTDefaultSettingsStorage.h"
 #import "OCTDefaultFileStorage.h"
 
-static NSString *const kDefaultSettingsStorageUserDefaultsKey = @"me.dvor.objcTox.settings";
 static NSString *const kDefaultBaseDirectory = @"me.dvor.objcTox";
 
 @implementation OCTManagerConfiguration
@@ -20,9 +18,6 @@ static NSString *const kDefaultBaseDirectory = @"me.dvor.objcTox";
 + (instancetype)defaultConfiguration
 {
     OCTManagerConfiguration *configuration = [OCTManagerConfiguration new];
-
-    configuration.settingsStorage = [[OCTDefaultSettingsStorage alloc]
-                                     initWithUserDefaultsKey:kDefaultSettingsStorageUserDefaultsKey];
 
     NSString *baseDirectory = [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES) firstObject];
     baseDirectory = [baseDirectory stringByAppendingPathComponent:kDefaultBaseDirectory];
@@ -45,6 +40,8 @@ static NSString *const kDefaultBaseDirectory = @"me.dvor.objcTox";
     configuration.options.proxyPort = 0;
     configuration.options.tcpPort = 0;
 
+    configuration.importToxSaveFromPath = nil;
+
     return configuration;
 }
 
@@ -54,9 +51,10 @@ static NSString *const kDefaultBaseDirectory = @"me.dvor.objcTox";
 {
     OCTManagerConfiguration *configuration = [[[self class] allocWithZone:zone] init];
 
-    configuration.settingsStorage = self.settingsStorage;
     configuration.fileStorage = self.fileStorage;
     configuration.options = [self.options copy];
+    configuration.importToxSaveFromPath = [self.importToxSaveFromPath copy];
+    configuration.passphrase = [self.passphrase copy];
 
     return configuration;
 }