+++ Classes/Private/Manager/Database/OCTRealmManager.m
@@ -18,12 +18,14 @@
 #import "OCTMessageText.h"
 #import "OCTMessageFile.h"
 #import "OCTMessageCall.h"
+#import "OCTSettingsStorageObject.h"
 
 #import "DDLog.h"
 #undef LOG_LEVEL_DEF
 #define LOG_LEVEL_DEF LOG_LEVEL_VERBOSE
 
-static const uint64_t kCurrentSchemeVersion = 1;
+static const uint64_t kCurrentSchemeVersion = 2;
+static NSString *kSettingsStorageObjectPrimaryKey = @"kSettingsStorageObjectPrimaryKey";
 
 @interface OCTRealmManager ()
 
@@ -33,20 +35,7 @@ static const uint64_t kCurrentSchemeVersion = 1;
 @end
 
 @implementation OCTRealmManager
-
-#pragma mark -  Class methods
-
-+ (NSMutableSet *)migratedPathesSet
-{
-    static NSMutableSet *dictionary;
-    static dispatch_once_t token;
-
-    dispatch_once(&token, ^{
-        dictionary = [NSMutableSet new];
-    });
-
-    return dictionary;
-}
+@synthesize settingsStorage = _settingsStorage;
 
 #pragma mark -  Lifecycle
 
@@ -68,16 +57,8 @@ static const uint64_t kCurrentSchemeVersion = 1;
     dispatch_sync(_queue, ^{
         __strong OCTRealmManager *strongSelf = weakSelf;
 
-        @synchronized([OCTRealmManager class]) {
-            NSMutableSet *migratedPathesSet = [OCTRealmManager migratedPathesSet];
-
-            if (! [migratedPathesSet containsObject:path]) {
-                [migratedPathesSet addObject:path];
-                [strongSelf performMigrationForRealmWithPath:path];
-            }
-        }
-
-        strongSelf->_realm = [RLMRealm realmWithPath:path];
+        [strongSelf createRealmWithPath:path];
+        [strongSelf createSettingsStorage];
     });
 
     [self convertAllCallsToMessages];
@@ -188,6 +169,36 @@ static const uint64_t kCurrentSchemeVersion = 1;
 }
 
 #pragma mark -  Other methods
+
+- (void)createRealmWithPath:(NSString *)path
+{
+    RLMRealmConfiguration *configuration = [RLMRealmConfiguration defaultConfiguration];
+    configuration.path = path;
+    configuration.schemaVersion = kCurrentSchemeVersion;
+    configuration.migrationBlock = [self realmMigrationBlock];
+
+    NSError *error;
+    self->_realm = [RLMRealm realmWithConfiguration:configuration error:&error];
+
+    if (! self->_realm) {
+        DDLogInfo(@"OCTRealmManager: init failed with error %@", error);
+    }
+}
+
+- (void)createSettingsStorage
+{
+    _settingsStorage = [OCTSettingsStorageObject objectInRealm:self.realm
+                                                 forPrimaryKey:kSettingsStorageObjectPrimaryKey];
+
+    if (! _settingsStorage) {
+        DDLogInfo(@"OCTRealmManager: no _settingsStorage, creating it");
+        _settingsStorage = [OCTSettingsStorageObject new];
+
+        [self.realm beginWriteTransaction];
+        [self.realm addObject:_settingsStorage];
+        [self.realm commitWriteTransaction];
+    }
+}
 
 - (OCTFriend *)friendWithFriendNumber:(OCTToxFriendNumber)friendNumber
 {
@@ -388,14 +399,17 @@ static const uint64_t kCurrentSchemeVersion = 1;
 
 #pragma mark -  Private
 
-- (void)performMigrationForRealmWithPath:(NSString *)path
+- (RLMMigrationBlock)realmMigrationBlock
 {
-    [RLMRealm setSchemaVersion:kCurrentSchemeVersion forRealmAtPath:path withMigrationBlock:
-     ^(RLMMigration *migration, uint64_t oldSchemaVersion) {
-        if (oldSchemaVersion < 1) {
-            // objcTox version 0.1.0
-        }
-    }];
+    return ^(RLMMigration *migration, uint64_t oldSchemaVersion) {
+               if (oldSchemaVersion < 1) {
+                   // objcTox version 0.1.0
+               }
+
+               if (oldSchemaVersion < 2) {
+                   // objcTox version 0.2.1
+               }
+    };
 }
 
 - (RBQRealmChangeLogger *)logger