+++ Classes/Public/Manager/Configuration/OCTManagerConfiguration.h
@@ -8,7 +8,6 @@
 
 #import <Foundation/Foundation.h>
 
-#import "OCTSettingsStorageProtocol.h"
 #import "OCTFileStorageProtocol.h"
 #import "OCTToxOptions.h"
 
@@ -18,43 +17,53 @@
 @interface OCTManagerConfiguration : NSObject <NSCopying>
 
 /**
- * Settings storage to be used.
+ * File storage to use.
  *
- * By default OCTDefaultSettingsStorage will be used.
- */
-@property (strong, nonatomic) id<OCTSettingsStorageProtocol> settingsStorage;
-
-/**
- * File storage to be used.
- *
- * By default OCTDefaultFileStorage will be used.
- */
-@property (strong, nonatomic) id<OCTFileStorageProtocol> fileStorage;
-
-/**
- * Options for tox to use.
- */
-@property (strong, nonatomic) OCTToxOptions *options;
-
-/**
- * This is default configuration for manager. Parameters are follows
- *
- * - settings are stored in NSDictionary in NSUserDefaults for "me.dvor.objcTox.settings" key;
- *
+ * Default values: OCTDefaultFileStorage will be used with following parameters:
  * - tox save file is stored at "{app document directory}/me.dvor.objcTox/save.tox"
  * - downloaded files are stored at "{app document directory}/me.dvor.objcTox/downloads"
  * - uploaded files are stored at "{app document directory}/me.dvor.objcTox/uploads"
  * - avatars are stored at "{app document directory}/me.dvor.objcTox/avatars"
  * - temporary files are stored at NSTemporaryDirectory()
+ */
+@property (strong, nonatomic, nonnull) id<OCTFileStorageProtocol> fileStorage;
+
+/**
+ * Options for tox to use.
  *
+ * Default values:
  * - IPv6 support enabled
  * - UDP support enabled
  * - No proxy is used.
  *
- * @return Default configuration for OCTManager.
- *
  * @warning On mobile devices you may want to turn off UDP support to increase battery life.
  */
-+ (instancetype)defaultConfiguration;
+@property (strong, nonatomic, nonnull) OCTToxOptions *options;
+
+/**
+ * Passphrase used to decrypt tox save file.
+ * You should specify this parameter *only* if tox save file is already encrypted. If you would like to
+ * enable encryption please use OCTManager's method.
+ *
+ * Default value: nil.
+ */
+@property (strong, nonatomic, nullable) NSString *passphrase;
+
+/**
+ * If this parameter is set, tox save file will be copied from given path.
+ * You can set this property to import tox save from some other location.
+ *
+ * Default value: nil.
+ */
+@property (strong, nonatomic, nullable) NSString *importToxSaveFromPath;
+
+/**
+ * This is default configuration for manager.
+ * Each property of OCTManagerConfiguration has "Default value" field. This method returns configuration
+ * with those default values set.
+ *
+ * @return Default configuration for OCTManager.
+ */
++ (nonnull instancetype)defaultConfiguration;
 
 @end