+++ Tests/OCTRealmTests.m
@@ -10,31 +10,28 @@
 
 #import "OCTRealmTests.h"
 
+@interface OCTRealmTests ()
+
+@property (strong, nonatomic) id realmMock;
+
+@end
+
 @implementation OCTRealmTests
-
-- (NSString *)realmPath
-{
-    NSString *directory = [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES) firstObject];
-
-    return [directory stringByAppendingPathComponent:@"test.realm"];
-}
 
 - (void)setUp
 {
     [super setUp];
     // Put setup code here. This method is called before the invocation of each test method in the class.
 
-    NSString *realmPath = [self realmPath];
-    NSString *directory = [realmPath stringByDeletingLastPathComponent];
+    RLMRealmConfiguration *configuration = [RLMRealmConfiguration defaultConfiguration];
+    configuration.inMemoryIdentifier = @"OCTRealmTests";
 
-    NSFileManager *fileManager = [NSFileManager defaultManager];
-    if (! [fileManager fileExistsAtPath:directory]) {
-        // This is hack to fix xctool issue on Travis CI.
-        // For some reason it don't create documents directory.
-        [fileManager createDirectoryAtPath:directory withIntermediateDirectories:YES attributes:nil error:nil];
-    }
+    RLMRealm *realRealm = [RLMRealm realmWithConfiguration:configuration error:nil];
 
-    self.realmManager = [[OCTRealmManager alloc] initWithDatabasePath:realmPath];
+    self.realmMock = OCMClassMock([RLMRealm class]);
+    OCMStub([self.realmMock realmWithConfiguration:[OCMArg any] error:[OCMArg anyObjectRef]]).andReturn(realRealm);
+
+    self.realmManager = [[OCTRealmManager alloc] initWithDatabasePath:@"path"];
     self.realmManager = OCMPartialMock(self.realmManager);
 }
 
@@ -47,13 +44,8 @@
     [(id)self.realmManager stopMocking];
     self.realmManager = nil;
 
-    NSString *realmPath = [self realmPath];
-    NSString *lockPath = [realmPath stringByAppendingString:@".lock"];
-
-    NSFileManager *fileManager = [NSFileManager defaultManager];
-    [fileManager removeItemAtPath:realmPath error:nil];
-    [fileManager removeItemAtPath:lockPath error:nil];
-
+    [self.realmMock stopMocking];
+    self.realmMock = nil;
 
     // Put teardown code here. This method is called after the invocation of each test method in the class.
     [super tearDown];